+++ packages/graphql-codegen-zod/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @use-pico/graphql-codegen-zod
 
+## 4.0.5
+
+### Patch Changes
+
+- Fixed also input types
+
+## 4.0.4
+
+### Patch Changes
+
+- Fixed schema names
+
+## 4.0.3
+
+### Patch Changes
+
+- Improved generator using Zod v4 concepts
+
 ## 4.0.2
 
 ### Patch Changes